user and  auth  prisma

+++ src/app/api/todos/route.ts
@@ -1,6 +1,7 @@
 import { NextResponse, NextRequest } from 'next/server'
 import prisma from '../../../lib/prisma'
 import * as yup from 'yup';
+import { getUserServerSession } from '@/app/auth/actions/auth-actions';
 
 export async function GET(request: Request) { 
 
@@ -28,11 +29,17 @@ const postSchema = yup.object({
 
 export async function POST(request: Request) { 
 
+  const user = await getUserServerSession();
+
+  if ( !user ) {
+    return NextResponse.json('Unauthorized', { status: 401 });
+  }
+
   try {
     const {complete, description} = await postSchema.validate( await request.json() );
   
     const todo = await prisma.todo.create({
-      data: {complete, description}
+      data: {complete, description, userId: user.id}
     });
   
     return NextResponse.json(todo);
@@ -46,9 +53,15 @@ export async function POST(request: Request) {
 
 export async function DELETE(request: Request) { 
 
+  const user = await getUserServerSession();
+
+  if ( !user ) {
+    return NextResponse.json('Unauthorized', { status: 401 });
+  }
+
   try {  
     
-    await prisma.todo.deleteMany({ where: { complete: true } } );
+    await prisma.todo.deleteMany({ where: { complete: true, userId: user.id } } );
   
     return NextResponse.json('Ok delete all completed todos');
